Function to change one cell value. Changed name of EmptyCells to CellFormulas and moved some code over to file.
